Controller Terminé

+++ ConsoleApp1/Model/Staff/Chef.cs
@@ -9,8 +9,8 @@ namespace KitchenProject.Model.Staff
 {
     class Chef
     {
-        public float xPositionInit = 100;
-        public float yPositionInit = 100;
+        public int xPositionInit = 100;
+        public int yPositionInit = 100;
         public int widthInit = 100;
         public int heightInit = 100;
         static Image imageChef = Image.FromFile("D:/POO/CSHARP/RtrCsharp/asset/Staff/chef.png");
